Updated DB, applicant MVC, added Main Controller, views.

+++ src/main/java/Application/DataModel/ApplicantRepo.java
@@ -4,5 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ApplicantRepo extends CrudRepository<Applicant, Integer> {
+public interface ApplicantRepo extends CrudRepository<Applicant, Long> {
 }
+
